+++ nanobot/agent/autocompact.py
